AI in Action/server/requirements.txt
protobuf~=6.31.1
python-dotenv~=1.1.0
pydantic~=2.11.5
pandas~=2.3.0
pymongo~=4.13.1
uvicorn~=0.34.3
fastapi~=0.115.12